+++ clients/algoliasearch-client-java/CHANGELOG.md
@@ -1,3 +1,13 @@
+## [4.15.0](https://github.com/algolia/algoliasearch-client-java/compare/4.14.0...4.15.0)
+
+- [c3addb449](https://github.com/algolia/api-clients-automation/commit/c3addb449) feat(specs): add CT Query predicate property ([#4654](https://github.com/algolia/api-clients-automation/pull/4654)) by [@morganleroi](https://github.com/morganleroi/)
+
+## [4.14.0](https://github.com/algolia/algoliasearch-client-java/compare/4.13.0...4.14.0)
+
+- [03a7074763](https://github.com/algolia/api-clients-automation/commit/03a7074763) chore(deps): dependencies 2025-03-10 ([#4566](https://github.com/algolia/api-clients-automation/pull/4566)) by [@algolia-bot](https://github.com/algolia-bot/)
+- [26457d0ec8](https://github.com/algolia/api-clients-automation/commit/26457d0ec8) feat(specs): add the owner attributes to ingestion resources ([#4616](https://github.com/algolia/api-clients-automation/pull/4616)) by [@millotp](https://github.com/millotp/)
+- [cdf605e369](https://github.com/algolia/api-clients-automation/commit/cdf605e369) chore(deps): dependencies 2025-03-24 ([#4617](https://github.com/algolia/api-clients-automation/pull/4617)) by [@algolia-bot](https://github.com/algolia-bot/)
+
 ## [4.13.0](https://github.com/algolia/algoliasearch-client-java/compare/4.12.3...4.13.0)
 
 - [63a50acb3](https://github.com/algolia/api-clients-automation/commit/63a50acb3) docs(java): fix typo on README ([#4564](https://github.com/algolia/api-clients-automation/pull/4564)) by [@Giglium](https://github.com/Giglium/)